Home Page
I started developing UI for home page. The initial components are created. I need to restructure the grid of items to be its own component that dynamically changes size based on window size. I also made the DB. I need to make the views that read the DB. Also Stored procedures for updates and inserts. The application layer also needs to get built.